requirement.txt
absl-py==2.2.2
backcall==0.2.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cloudpickle==3.1.1
contourpy==1.3.2
crcmod==1.7
cycler==0.12.1
decorator==5.2.1
Farama-Notifications==0.0.4
filelock==3.18.0
fonttools==4.57.0
fsspec==2025.3.2
getmac==0.9.5
grpcio==1.59.3
gym==0.26.2
gym-notices==0.0.8
gymnasium==1.0.0
idna==3.10
imageio==2.37.0
ipython==7.31.1
jedi==0.17.2
Jinja2==3.1.6
kiwisolver==1.4.8
Markdown==3.8
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mininet==2.3.1b4
mpmath==1.3.0
multiagent==0.0.1
netifaces==0.11.0
networkx==3.4.2
nnpy==1.4.2
numpy==2.2.3
numpy-stl==3.2.0
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
p4runtime-shell==0.0.3.post21+g7374d7c.d20250415
packaging==24.2
pandas==2.2.3
parso==0.7.1
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.2.1
pip==24.0
ply==3.11
progressbar33==2.4
prompt_toolkit==3.0.51
protobuf==4.21.12
psutil==7.0.0
ptf==0.10.0.post15+ge29ef7e
ptyprocess==0.7.0
pycparser==2.22
pyglet==2.1.3
Pygments==2.19.1
pyparsing==3.2.3
pyperclip==1.8.2
python-dateutil==2.9.0.post0
python-utils==3.9.1
pytz==2025.2
requests==2.32.3
scapy==2.5.0
scapy-helper==0.14.8
scipy==1.15.2
setuptools==78.1.0
six==1.17.0
sympy==1.13.1
tabulate==0.8.10
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
thrift==0.16.0
torch==2.5.1
traitlets==5.14.3
triton==3.1.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.45.1
